split resource usage logger
